Anymize-UI-1.1/anymize/requirements.txt
# Core Web Framework
Flask==2.0.1

# HTTP Requests
requests==2.28.2

# Utility progress bar
tqdm

# OCR & Document Processing
pytesseract==0.3.13
easyocr==1.7.2
docx2txt==0.8
tika==2.6.0
pdf2image==1.17.0
PyMuPDF==1.22.5

# Image Processing
Pillow==9.5.0
opencv-python==4.7.0.72

# General Utilities
numpy==1.24.2
json5
Faker
gunicorn
